Got tired of wrestling with it, PyOOP FTW.
I just couldn't quite wrap my head around the subtlies of the hardware
RISC architecture so I "cheated" and made a simple Python model of it
instead. This is unsatifactory in the sense that the design of the RISC
is really elegant and interesting while the pythonic model cuts corners.
For example, the stall state-machine for handling memory instructions on
the VN architecture is totally given a miss.  The system just has a
reference to its ram (a dict or dict-like object) and accesses it as it
pleases.

Since this model is only intended to allow me to play with the Oberon
system on my laptop, and since I could get the verilog sources running in
a sim or just purchase an FPGA board, I don't feel too badly about it.
Still, it is disappointing to discover that my brain ain't what it used
to be: I can appreciate the chip without being quite able to grok it
(yet.)