Trap on BR to Module Table.
The first entry in the MT should be an instruction to jump to a trap
routine, instead for the moment we simply check for jumps to the MT first
entry and report them as a Python exception.  I need to go back and
modify my bit of reverse-engineered Modules.py to start module numbers
(offsets into the MT) at 1 rather than zero.  It might not make sense to
actually stick a trap handler in there until I have a trap handler
routine to stick. ;-)