I guess you don't need that first noop?
I must have put it there for a reason?  Oh yeah!  The Joy interpreter I
was working on used the Zeroth word as the root of all stacks (or
something like that, it's been over a year) so it had to point to
itself, or something.  Anyway, that's enough information, however dimly
recalled, to put my concern about that no-op to rest.  Whew!